refactor(server): S4-06 — Extract Module for NLOC compliance
Extract async worker wrappers to workers.rs (pipeline.rs 599->536 NLOC).
Extract apply_advanced_with_rollback + log_rollback_invariant from
create_vector_collection (CC 11->7). Both changes follow Fowler's
Extract Module pattern.
