+++ crates/velesdb-server/src/handlers/search/multi.rs
@@ -12,8 +12,9 @@ use crate::types::{ErrorResponse, MultiQuerySearchRequest, SearchResponse};
 use crate::AppState;
 
 use super::pipeline::{
-    finish_search_with_cb, parse_filter_or_400, run_blocking_search, validate_query_dimension,
+    finish_search_with_cb, parse_filter_or_400, validate_query_dimension,
 };
+use super::workers::run_blocking_search;
 use crate::handlers::helpers::{apply_pre_check, extract_client_id, get_vector_collection_or_404};
 
 /// Multi-query search with fusion strategies.
